+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Made save button sticky.
 * Fixed error messages.
+* Added call to middleware to get which channels are in the middleware
+  to make sure it matches the backend.
 * Added watch parameter to datetime pickers.
 * Changed blameable user field from string to reference.